security-service/requirements.txt
Flask
flask-cors
requests
gunicorn
pyjwt
attrs